+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - Criterions
     - [**NEW**] _Manual_
         - Criterion 'manual' requests user to specify fitness value for each candidate.
+- Interpolators
+    - [**NEW**] _None_
+        - Dummy interpolator that forwards given points.
 - Penalizers
     - [**NEW**] _None_
         - Dummy penalizer that allows any candidate.